Rearrange a number of functions to remove cyclic dependency
